Answering Self-Study Question 3 - What does CRUD stand for?
* (C)reate
* (R)etrieve
* (U)date
* (D)elete

Web Developers that are familiar with REST term, can think of CRUD as the original that REST (Representational State Transfer) was derived from. It is a type of order of operations to help understand the functions that almost any database type application should have. (Meaning any application that pulls and retrieves lists of data to be displayed.)